% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gst_catalogue}
\alias{gst_catalogue}
\title{GST Data Catalogue}
\usage{
gst_catalogue()
}
\value{
Invisibly returns a \code{data.table} of dataset metadata.
}
\description{
Print a summary of all datasets available in the package.
}
\examples{
gst_catalogue()
}
