% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gst_registration}
\alias{gst_registration}
\title{State-wise GST Taxpayer Registration (as of March 2025)}
\format{
A \code{data.table} with 39 rows and 9 columns:
\describe{
\item{state_code}{\code{character}. GST state code.}
\item{state}{\code{character}. State/UT name.}
\item{region}{\code{character}. Geographic region.}
\item{normal}{\code{numeric}. Normal (regular) taxpayers.}
\item{composition}{\code{numeric}. Composition scheme taxpayers.}
\item{isd}{\code{numeric}. Input Service Distributors.}
\item{casual}{\code{numeric}. Casual taxpayers.}
\item{tcs}{\code{numeric}. Tax Collectors at Source.}
\item{tds}{\code{numeric}. Tax Deductors at Source.}
}
}
\source{
GST Council / GSTN. \url{https://www.gst.gov.in}
}
\usage{
gst_registration
}
\description{
Number of registered GST taxpayers per state/UT, categorised by
registration type. Snapshot as of 31 March 2025.
}
\examples{
data(gst_registration)
gst_registration[order(-normal), .(state, normal, composition)][1:10]
}
\keyword{datasets}
